dom/workers/test/test_bug1132395.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test for 1132395</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<script class="testbody" type="text/javascript">

// This test is full of dummy debug messages. This is because I need to follow
// an hard-to-reproduce timeout failure.

info("test started");
SpecialPowers.pushPrefEnv({ set: [["dom.workers.sharedWorkers.enabled", true]] }, function() {
  info("test started");
  var sw = new SharedWorker('bug1132395_sharedWorker.js');
  sw.port.onmessage = function(event) {
    info("sw.onmessage received");
    ok(true, "We didn't crash.");
    SimpleTest.finish();
  }

  info("sw.postmessage called");
  sw.port.postMessage('go');
});

SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
